requirements.txt
opencv-python
pillow
picamera2